requirements.txt
yfinance>=0.2.28
pandas>=2.1.4
numpy>=1.26.0
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0
streamlit>=1.29.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
selenium>=4.15.0
webdriver-manager>=4.0.1
mcp>=1.0.0
psutil>=5.9.0
nltk>=3.9.1
websockets>=11.0.3
